requirements.txt
playwright==1.52.0
pytest==8.4.1
pytest-playwright==0.7.1
allure-pytest==2.15.0
python-dotenv==1.1.1
PyYAML==6.0.2
